requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Dance==7.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
URLObject==2.4.3
Werkzeug==3.1.3
